Move activities stores to component folders
