+++ src/scenes/Activities/ActivityHistory/store.js
@@ -1,11 +1,10 @@
 import { observable, action, reaction, computed } from 'mobx';
-import { TransactionType, SortBy, Routes } from 'constants';
 import _ from 'lodash';
+import { TransactionType, SortBy, Routes } from 'constants';
+import { Transaction, Oracle } from 'models';
 
-import { getDetailPagePath } from '../../helpers/utility';
-import { queryAllTransactions, queryAllOracles } from '../../network/graphQuery';
-import Transaction from '../models/Transaction';
-import Oracle from '../models/Oracle';
+import { getDetailPagePath } from '../../../helpers/utility';
+import { queryAllTransactions, queryAllOracles } from '../../../network/graphQuery';
 
 const INIT_VALUES = {
   transactions: [],
