+++ src/scenes/Activities/ActivityHistory/styles.js
@@ -1,0 +1,7 @@
+const styles = () => ({
+  historyTable: {
+    overflowX: 'scroll',
+  },
+});
+
+export default styles;
